fix train and recognize logic
