+++ opencv/opencv/opencv.vcxproj.filters
@@ -23,7 +23,7 @@
     <ClInclude Include="stb_image.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="stb_image_resize2.h">
+    <ClInclude Include="stb_image_resize.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
